4.21*: bloglist expansion, step9

+++ controllers/blogs.js
@@ -55,8 +55,36 @@ blogsRouter.put('/:id', async (req, res) => {
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
+	// get token
+	const decodedToken = jwt.verify(req.token, process.env.SECRET)
+
+	// check if token valid
+	if (!req.token || !decodedToken.id) 
+		return res
+			.status(401)
+			.json({ error: 'missing or invalid token' })
+	
+	const user = await User.findById(decodedToken.id)
+
+	// get blog by id
+	const blogToDelete = await Blog.findById(req.params.id)
+
+	if (!blogToDelete)
+		return res
+			.status(404)
+			.end()
+	
+	// check if the logged-in user is the same as the blog's author
+	if (blogToDelete.author.toString() !== user.id.toString()) {
+		return res
+			.status(401)
+			.json({ error: 'access denied' })
+	}
+
 	await Blog.findByIdAndRemove(req.params.id)
-	res.status(204).end()
+	res
+		.status(204)
+		.end()
 })
 
 module.exports = blogsRouter
